requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
Pillow==9.2.0
python-dotenv==0.20.0
python-slugify==6.1.2
sqlparse==0.4.2
text-unidecode==1.3
